updated rewards

+++ src/StepResult.java
@@ -1,9 +1,9 @@
 public class StepResult {
     public String obs;
-    public float reward;
+    public double reward;
     public boolean done;
 
-    public StepResult(String obs, float reward, boolean done) {
+    public StepResult(String obs, double reward, boolean done) {
         this.obs = obs;
         this.reward = reward;
         this.done = done;
